Moved jordattebayo contents to root of folder
