+++ components/globalstyles.tsx
