--- jordattebayo/.gitignore
@@ -1,32 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-*.txt
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# next.js
-/.next/
-/out/
-
-# production
-/build
-
-# misc
-.DS_Store
-
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# local env files
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
