+++ pages/api/projects/[slug].ts
